% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{ComNet_data}
\alias{ComNet_data}
\title{Toy example ComNet dataset}
\format{A ComNet object.}
\usage{
ComNet_data
}
\description{
An example dataset comprising 20 messages sent between 10 actors.
}
\keyword{datasets}

