% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_output.R
\name{generate_output}
\alias{generate_output}
\title{Generate Interpretable Output from a CCAS object.}
\usage{
generate_output(CCAS_Object, output_directory = NULL,
  output_name_stem = NULL, generate_plots = TRUE)
}
\arguments{
\item{CCAS_Object}{The object returned by the ccas() main estimation
function.}

\item{output_directory}{Defaults to NULL, currently unimplemented.}

\item{output_name_stem}{Defaults to NULL, currently unimplemented.}

\item{generate_plots}{Logical indicating whether plots should be generated,
defaults to TRUE, but may be set to FALSE if the user only wishes to access
aggregate data.}
}
\value{
A series of plots and a CCAS_Object containing useful output.
}
\description{
Generates a large number of plots of various posterior quantites
of interest, includingtrace plots, parameter estimates and top words.
}

