% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiR_PP_Plots.R
\name{GiR_PP_Plots}
\alias{GiR_PP_Plots}
\title{Make Nice GiR PP-Plots}
\usage{
GiR_PP_Plots(forward_samples, backward_samples, generate_PP_plot = TRUE)
}
\arguments{
\item{forward_samples}{A dataframe with statistic generated by
test_internal_functions() with Getting_It_Right = TRUE.}

\item{backward_samples}{A dataframe with statistic generated by
test_internal_functions() with Getting_It_Right = TRUE.}

\item{generate_PP_plot}{Defaults to TRUE. If FALSE, then QQ plots are
generated.}
}
\value{
A bunch of plots.
}
\description{
Makes Nice PP Plots with Getting it Right Results.
}
