% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{Model_Output}
\alias{Model_Output}
\title{Department Manager Email dataset -- Example output from the ccas() function.}
\format{A CCAS object.}
\usage{
Model_Output
}
\description{
Output from the example model run on county manager data..
}
\keyword{datasets}
