% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_pattern_log_likelihood.R
\name{plot_interaction_pattern_log_likelihood}
\alias{plot_interaction_pattern_log_likelihood}
\title{Trace plot of interaction pattern model log likelihood}
\usage{
plot_interaction_pattern_log_likelihood(CCAS_Object)
}
\arguments{
\item{CCAS_Object}{devtoolAn object of class CCAS containing estimation results.}
}
\value{
A plot.
}
\description{
Generates a trace plot of the interaction pattern model log
likelihood. This is a potential way to diagnose the convergence of the model.
}
\examples{
\dontrun{
# load in saved model output from ccas() function.
data(Model_Output)
plot_interaction_pattern_log_likelihood(Model_Output)
}
}

