% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_pattern_parameter_estimates.R
\name{plot_interaction_pattern_parameter_estimates}
\alias{plot_interaction_pattern_parameter_estimates}
\title{Plot interaction-pattern sepcific coefficients.}
\usage{
plot_interaction_pattern_parameter_estimates(CCAS_Object,
  interaction_pattern_index, coefficient_names = NULL,
  leave_out_coefficients = NULL, normalize_coefficients = FALSE,
  generate_plot = TRUE)
}
\arguments{
\item{CCAS_Object}{The object returned by the ccas() main estimation
function.}

\item{interaction_pattern_index}{The index of the cluster specific
communication network we wish to plot.}

\item{coefficient_names}{Defualts to NULL. Can be a string vector of names
for coefficients to be used in making publication quality plots.}

\item{leave_out_coefficients}{Defaults to NULL. Can be a string vector of
coefficient names as they appear in the plot. These coefficients will be
removed from the final plot. Useful if the intercept term is much larger in
magnitude than other estimates, and the user wishes to clarify the other
parameter estimates without normalizing.}

\item{normalize_coefficients}{Defaults to FALSE, if TRUE then parameter
estimates will be converted be deivided by their standard deviations with
and displayed with 95 percent confidence intervals. These coefficients will
no longer be comparable, but make graphical interpretation of significance
and sign easier.}

\item{generate_plot}{Logical indicating whether a plot should be generated,
defaults to TRUE, but may be set to FALSE if the user only wishes to access
parameter estimates.}
}
\value{
A plot.
}
\description{
Plots parameter estimates and intercept term associated with a
particular interaction pattern.
}

