% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameter_estimates.R
\name{plot_parameter_estimates}
\alias{plot_parameter_estimates}
\title{Plot interaction-pattern sepcific sub-networks}
\usage{
plot_parameter_estimates(CCAS_Object, plots_per_row, generate_plots = TRUE)
}
\arguments{
\item{CCAS_Object}{The object returned by the ccas() main estimation
function.}

\item{plots_per_row}{The number of interaction pattern coefficient plots to
include in each row of the output figure.}

\item{generate_plots}{Logical indicating whether plots should be generated,
defaults to TRUE, but may be set to FALSE if the user only wishes to access
parameter estimates.}
}
\value{
A plot.
}
\description{
Plots coefficient ad intercept estimates for all interaction
patterns.
}
\examples{
\dontrun{
# load in saved model output from ccas() function.
data(Model_Output)
# plots_per_row should be equal to the number of interactions patterns.
summary_data <- plot_parameter_estimates(Model_Output, plots_per_row = 4)
}
}
