% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topic_model_log_likelihood.R
\name{plot_topic_model_log_likelihood}
\alias{plot_topic_model_log_likelihood}
\title{Trace plot of topic model log likelihood}
\usage{
plot_topic_model_log_likelihood(CCAS_Object, topic_model_burnin = NULL)
}
\arguments{
\item{topic_model_burnin}{The number of iterations of Gibbs sampling to
discard before generating trace plot and calculating Geweke statistic.}

\item{An}{object of class CCAS containing estimation results.}
}
\value{
A plot.
}
\description{
Generates a trace plot of the unnormalized topic model log
likelihood. This is a potential way to diagnose the convergence of the model.
}

