% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare communication network data for analysis.}
\usage{
prepare_data(document_authors, document_edge_matrix, document_term_matrix,
  covariate_data = NULL, vocabulary = NULL)
}
\arguments{
\item{document_authors}{A vector recording the index of the sender of each
document.}

\item{document_edge_matrix}{A documents x actors matrix recording whether
actor j was a recipient of message i.}

\item{document_term_matrix}{A documents x unique terms matrix recording the
count of unique term j in document i.}

\item{covariate_data}{Defaults to NULL.}

\item{vocabulary}{Defaults to NULL.}
}
\value{
An object of class ComNet
}
\description{
Reads in and stores communication network data in a ComNet
object for use in main estimation procedure.
}

