% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_Documentation.R
\docType{package}
\name{SpeedReader}
\alias{SpeedReader}
\alias{SpeedReader-package}
\title{CCAS: A Model for Communication Network Content and Structure}
\description{
CCAS: A Model for Communication Network Content and Structure
}
\section{CCAS functions}{

This package will eventually provide functions to estimate a statistical model
of communication network data including email and blog posts.
}

