% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_words.R
\name{top_words}
\alias{top_words}
\title{Generate Topic top words}
\usage{
top_words(CCAS_Object, output_LaTeX_code)
}
\arguments{
\item{CCAS_Object}{The object returned by the ccas() main estimation
function.}

\item{output_LaTeX_code}{Logical indicating whether LaTeX code should be
output by the functon for incorporating in tables.}
}
\value{
A list of data.frames with topic and interaction pattern top words.
}
\description{
Generates data.frames containing the top words associated with
each topic and with each interaction pattern
}

