% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccas.R
\name{ccas}
\alias{ccas}
\title{Infer Communication Content and Structure}
\usage{
ccas(formula, interaction_patterns = 4, topics = 40, alpha = 1,
  beta = 0.01, iterations = 1000, metropolis_hastings_iterations = 500,
  metropolis_hastings_burnin = 500, target_accept_rate = 0.25,
  tollerance = 0.05, LSM_intercept_prior_variance = 5,
  LSM_intercept_prior_mean = 0, LSM_position_prior_variance = 5,
  LSM_position_prior_mean = 0, LSM_coefficient_prior_variance = 5,
  LSM_coefficient_prior_mean = 0)
}
\arguments{
\item{formula}{A formula object of the form 'ComNet ~
euclidean(d = 2)' where d is the number of dimensions in the latent space
that the user would like to include, and Comnet is a ComNet object generated
by the prepare_data() function. May also include optional terms
'sender("covariate_name")', receiver("covariate_name")',
'nodemix("covariate_name", base = value)' and netcov("network_covariate"),
which are defined analogously to the arguments in the latentnet package.}

\item{interaction_patterns}{Defaults to 4.}

\item{topics}{Defaults to 40.}

\item{alpha}{Defaults to 1.}

\item{beta}{Defaults to 0.01.}

\item{iterations}{Defaults to 1000.}

\item{metropolis_hastings_iterations}{Defaults to 500.}

\item{metropolis_hastings_burnin}{Defaults to 500.}

\item{target_accept_rate}{Defaults to 0.25.}

\item{tollerance}{Defaults to 0.05.}

\item{LSM_intercept_prior_variance}{Defaults to 5.}

\item{LSM_intercept_prior_mean}{Defaults to 0.}

\item{LSM_position_prior_variance}{Defaults to 5.}

\item{LSM_position_prior_mean}{Defaults to 0.}

\item{LSM_coefficient_prior_variance}{Defaults to 5.}

\item{LSM_coefficient_prior_mean}{Defaults to 0.}
}
\value{
An object of class CCAS containing estimation results.
}
\description{
Performs inference on the content conditional structure of a
text valued communication network.
}

