% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccas.R
\name{ccas}
\alias{ccas}
\title{Infer Communication Content and Structure}
\usage{
ccas(specification, document_authors, document_edge_matrix,
  document_term_matrix, covariate_data = NULL, vocabulary = NULL,
  interaction_patterns = 4, topics = 40, alpha = 1, beta = 0.01,
  iterations = 1000, metropolis_hastings_iterations = 500,
  target_accept_rate = 0.25, tollerance = 0.05)
}
\arguments{
\item{specification}{A formulat object of the form '~ euclidean(d = 2)' where
d is the number of dimensions in the latent space that the user would like to
include. may also include optional terms 'sender("covariate_name")',
'receiver("covariate_name")' and 'nodemix("covariate_name", base = value)'
which are defined analogously to the arguments in the latentnet package.}

\item{document_authors}{A vector recording the index of the sender of each
document.}

\item{document_edge_matrix}{A documents x actors matrix recording whether
actor j was a recipient of message i.}

\item{document_term_matrix}{A documents x unique terms matrix recording the
count of unique term j in document i.}

\item{covariate_data}{Defaults to NULL.}

\item{vocabulary}{Defaults to NULL.}

\item{interaction_patterns}{Defaults to 4.}

\item{topics}{Defaults to 40.}

\item{alpha}{Defaults to 1.}

\item{beta}{Defaults to 0.01.}

\item{iterations}{Defaults to 1000.}

\item{metropolis_hastings_iterations}{Defaults to 500.}

\item{target_accept_rate}{Defaults to 0.25.}

\item{tollerance}{Defaults to 0.05.}
}
\value{
An object of class CCAS containing estimation results.
}
\description{
Performs inference on the content conditional structure of a
text valued communication network.
}

