% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Documentation.R
\docType{data}
\name{ComNet_data}
\alias{ComNet_data}
\title{Department Manager Email dataset --  ComNet object}
\format{A ComNet object.}
\usage{
ComNet_data
}
\description{
An example dataset comprising 121 emails sent among 20 county government
department managers.
}
\keyword{datasets}

