% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interaction_pattern_network.R
\name{plot_interaction_pattern_network}
\alias{plot_interaction_pattern_network}
\title{Plot interaction-pattern sepcific sub-networks}
\usage{
plot_interaction_pattern_network(CCAS_Object, interaction_pattern_index,
  plot_color_category = NULL, generate_plot = TRUE)
}
\arguments{
\item{CCAS_Object}{The object returned by the ccas() main estimation
function.}

\item{interaction_pattern_index}{The index of the cluster specific
communication network we wish to plot.}

\item{plot_color_category}{The name of a categorical covariate the user
wishes to color nodes by. For example, nodes might be colored by gender. This
function only currently supports plotting by a categroical variable
containing five or less unique values. Defaults to NULL, in which case all
nodes are plotted in the same color.}

\item{generate_plot}{Logical indicating whether plot should be printed.
Defaults to TRUE, but can be set to FALSE if the user only wishes to access
the mean latent positions of actors.}
}
\value{
An data.frame containing the mean latent coordinates of each actor.
}
\description{
Plots a sub network associated with a particular interaction
pattern with nodes placed according to their posterior mean latent positions.
}
\examples{
\dontrun{
# load in saved model output from ccas() function.
data(Model_Output)
mean_positions <- plot_interaction_pattern_network(Model_Output,
                                 interaction_pattern_index = 1,
                                 plot_color_category = "Gender")
}
}

