% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_internal_functions.R
\name{test_internal_functions}
\alias{test_internal_functions}
\title{Test C++ functions}
\usage{
test_internal_functions(Test_Log_Space_Multinomial_Sampler = FALSE,
  Test_Edge_Probability = FALSE, Test_Sum_Over_T_Edge_Probs = FALSE,
  Test_Prior_Pobability_Of_I_P_Params = FALSE,
  Test_Sample_New_I_P_Parameters = FALSE, Test_LSM_Contribution = FALSE,
  Test_LDA_Contribution = FALSE,
  Test_Update_Single_Token_Topic_Assignment = FALSE, envir = environment(),
  ...)
}
\arguments{
\item{Test_Log_Space_Multinomial_Sampler}{Defualts to FALSE. If TRUE, then
optional arguments 'distribution' and 'seed' must be provided.}

\item{Test_Edge_Probability}{Defaults to FALSE. If TRUE, then optional arguments
'intercepts', 'coefficients', 'latent_pos', 'sender', 'recipient',
'current_covariates', 'interaction_pattern', and 'using_coefficients' must be
provided.}

\item{Test_Sum_Over_T_Edge_Probs}{Defaults to FALSE. If TRUE, then optional
arguments 'edge_probs', 'tokens_in_document', 'current_token_topic_assignment',
'current_document_topic_counts', 'leave_out_current_token',
'topic_interaction_patterns', 'document_sender','document_recipient', and
'leave_out_topic'. must be provided.}

\item{Test_Prior_Pobability_Of_I_P_Params}{Defaults to FALSE. If TRUE, then optional
arguments intercepts, coefficients, latent_pos, intercept_prior_mean,
intercept_prior_variance, coefficient_prior_mean, coefficient_prior_variance,
latent_position_prior_mean, latent_position_prior_variance,
using_coefficients must be provided.}

\item{Test_Sample_New_I_P_Parameters}{Defaults to FALSE. If TRUE, then optional
arguments intercepts, coefficients, latent_pos, intercept_prior_variance,
coefficient_prior_variance, latent_position_prior_variance,
using_coefficients must be provided.}

\item{Test_LSM_Contribution}{Defaults to FALSE. If TRUE, then optional
arguments edge_probs, tokens_in_document, topic,
current_token_topic_assignment, current_document_topic_counts,
document_edge_values, topic_interaction_patterns, document_sender must be
 provided.}

\item{Test_LDA_Contribution}{Defaults to FALSE. If TRUE, then optional
arguments tokens_in_document, current_token_topic_assignment,
current_document_topic_counts, word_type_topic_counts, topic_token_counts,
topic, current_word_type, alpha_m, beta_n, beta  must be provided.}

\item{Test_Update_Single_Token_Topic_Assignment}{Defaults to FALSE. If TRUE,
then optional arguments edge_probs, tokens_in_document,
current_token_topic_assignment, current_document_topic_counts,
word_type_topic_counts, topic_token_counts, current_word_type, alpha_m,
beta_n, document_edge_values, topic_interaction_patterns, document_sender,
 rand_num must be provided.}

\item{envir}{Should not be changed by the user, captures the current
environment to facilitate testing.}

\item{...}{optional arguments necessary to run each of the internal functions.}
}
\value{
Whatever is returned by the internal function being tested
}
\description{
Allows the user to interface with the individual C++ functions.
}

