% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_internal_functions.R
\name{test_internal_functions}
\alias{test_internal_functions}
\title{Test C++ functions}
\usage{
test_internal_functions(Test_Log_Space_Multinomial_Sampler = FALSE, ...)
}
\arguments{
\item{Test_Log_Space_Multinomial_Sampler}{Defualts to FALSE. If TRUE, then
optional arguments 'distribution' and 'seed' must be provided.}

\item{...}{optional arguments necessary to run each of the internal functions.}
}
\value{
Whatever is returned by the internal function being tested
}
\description{
Allows the user to interface with the individual C++ functions.
}
\examples{
\dontrun{

}
}

