% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_internal_functions.R
\name{test_internal_functions}
\alias{test_internal_functions}
\title{Test C++ functions}
\usage{
test_internal_functions(Test_Log_Space_Multinomial_Sampler = FALSE,
  Test_Edge_Probability = FALSE, Test_Sum_Over_T_Edge_Probs = FALSE,
  Test_Prior_Pobability_Of_I_P_Params = FALSE,
  Test_Sample_New_I_P_Parameters = FALSE, Test_LSM_Contribution = FALSE,
  Test_LDA_Contribution = FALSE,
  Test_Update_Single_Token_Topic_Assignment = FALSE,
  Test_Update_All_Token_Topic_Assignments = FALSE,
  Test_Update_Interaction_Pattern_Parameters = FALSE,
  Test_Update_Topic_Interaction_Pattern_Assignments = FALSE,
  Test_Adaptive_Metropolis = FALSE, Test_Inference = FALSE,
  Test_RDirichlet = FALSE,
  Test_Sample_Token_Topics_From_Generative_Process = FALSE, seed = NULL,
  distribution = NULL, intercepts = NULL, coefficients = NULL,
  latent_positions = NULL, document_sender = NULL,
  document_recipient = NULL, current_covariates = NULL,
  interaction_pattern = NULL, using_coefficients = NULL,
  edge_probabilities = NULL, tokens_in_document = NULL,
  current_token_topic_assignment = NULL,
  current_document_topic_counts = NULL, leave_out_current_token = NULL,
  topic_interaction_patterns = NULL, leave_out_topic = NULL,
  intercept_prior_mean = NULL, intercept_prior_standard_deviation = NULL,
  intercept_proposal_standard_deviations = NULL,
  coefficient_prior_mean = NULL,
  coefficient_prior_standard_deviation = NULL,
  coefficient_proposal_standard_deviations = NULL,
  latent_position_prior_mean = NULL,
  latent_position_prior_standard_deviation = NULL,
  latent_position_proposal_standard_deviations = NULL, topic = NULL,
  document_edge_values = NULL, word_type_topic_counts = NULL,
  topic_token_counts = NULL, current_word_type = NULL, alpha_m = NULL,
  beta_n = NULL, beta = NULL, rand_num = NULL, author_indexes = NULL,
  document_edge_matrix = NULL, document_topic_counts = NULL,
  token_topic_assignments = NULL, token_word_types = NULL,
  covariates = NULL, random_numbers = NULL, accept_rates = NULL,
  target_accept_rate = NULL, tollerance = NULL, update_size = NULL,
  iterations = NULL, metropolis_iterations = NULL,
  total_number_of_tokens = NULL, iterations_before_t_i_p_updates = NULL,
  update_t_i_p_every_x_iterations = NULL,
  perform_adaptive_metropolis = NULL, resample_word_types = NULL,
  slice_sample_alpha_m = NULL, slice_sample_step_size = NULL,
  parallel = NULL, use_cached_token_topic_distribution = NULL,
  cached_token_topic_distribution = NULL)
}
\arguments{
\item{Test_Log_Space_Multinomial_Sampler}{Defualts to FALSE. If TRUE, then
optional arguments distribution and seed must be provided.}

\item{Test_Edge_Probability}{Defaults to FALSE. If TRUE, then optional arguments
intercepts, coefficients, latent_pos, document_sender, document_recipient,
current_covariates, interaction_pattern, and using_coefficients must be
provided.}

\item{Test_Sum_Over_T_Edge_Probs}{Defaults to FALSE. If TRUE, then optional
arguments edge_probs, tokens_in_document, current_token_topic_assignment,
current_document_topic_counts, leave_out_current_token,
topic_interaction_patterns, document_sender,document_recipient,
leave_out_topic must be provided.}

\item{Test_Prior_Pobability_Of_I_P_Params}{Defaults to FALSE. If TRUE, then
optional arguments intercepts, coefficients, latent_pos, intercept_prior_mean,
intercept_prior_standard_deviation, coefficient_prior_mean,
coefficient_prior_standard_deviation, latent_position_prior_mean,
latent_position_prior_standard_deviation, using_coefficients must be
provided.}

\item{Test_Sample_New_I_P_Parameters}{Defaults to FALSE. If TRUE, then
optional  arguments intercepts, coefficients, latent_pos,
intercept_proposal_standard_deviations, coefficient_proposal_standard_deviations,
latent_position_proposal_standard_deviations, (all of which must be of
length = number of interaction patterns, and should all be equal)
using_coefficients must be provided.}

\item{Test_LSM_Contribution}{Defaults to FALSE. If TRUE, then optional
arguments edge_probs, tokens_in_document, topic,
current_token_topic_assignment, current_document_topic_counts,
document_edge_values, topic_interaction_patterns, document_sender must be
 provided.}

\item{Test_LDA_Contribution}{Defaults to FALSE. If TRUE, then optional
arguments tokens_in_document, current_token_topic_assignment,
current_document_topic_counts, word_type_topic_counts, topic_token_counts,
topic, current_word_type, alpha_m, beta_n, beta  must be provided.}

\item{Test_Update_Single_Token_Topic_Assignment}{Defaults to FALSE. If TRUE,
then optional arguments edge_probs, tokens_in_document,
current_token_topic_assignment, current_document_topic_counts,
word_type_topic_counts, topic_token_counts, current_word_type, alpha_m,
beta_n, document_edge_values, topic_interaction_patterns, document_sender,
rand_num must be provided.}

\item{Test_Update_All_Token_Topic_Assignments}{Defaults to FALSE. If TRUE,
then optional arguments author_indexes, document_edge_matrix,
topic_interaction_patterns, document_topic_counts, word_type_topic_counts,
topic_token_counts, token_topic_assignments, token_word_types, intercepts,
coefficients, latent_pos, covars, alpha_m, beta_n, random_numbers,
using_coefficients must be provided. Make sure that random_numbers has length
equal to the total number of tokens in the corpus.}

\item{Test_Update_Interaction_Pattern_Parameters}{Defaults to FALSE. If TRUE,
then optional arguments author_indexes, document_edge_matrix,
document_topic_counts, topic_interaction_patterns, intercepts, coefficients,
latent_pos, covars, using_coefficients, intercept_prior_mean,
intercept_prior_standard_deviation, intercept_proposal_standard_deviations,
coefficient_prior_mean, coefficient_prior_standard_deviation,
coefficient_proposal_standard_deviations, latent_position_prior_mean,
latent_position_prior_standard_deviation,
latent_position_proposal_standard_deviations, rand_num, edge_probs must be
provided.}

\item{Test_Update_Topic_Interaction_Pattern_Assignments}{Defaults to FALSE.
If TRUE, then optional arguments author_indexes, document_edge_matrix,
document_topic_counts, topic_interaction_patterns, intercepts, coefficients,
latent_pos, covars, using_coefficients, random_numbers, edge_probs must be
provided.}

\item{Test_Adaptive_Metropolis}{Defaults to FALSE. If TRUE, then optional
arguments intercept_proposal_standard_deviations,
coefficient_proposal_standard_deviations,
latent_position_proposal_standard_deviations,
accept_rates, target_accept_rate,
tollerance, update_size must be provided.}

\item{Test_Inference}{If TRUE, then optional arguments author_indexes,
document_edge_matrix, document_topic_counts, topic_interaction_patterns,
word_type_topic_counts, topic_token_counts, token_topic_assignments,
token_word_types, intercepts, coefficients, latent_pos, covars, alpha_m,
beta_n, using_coefficients, intercept_prior_mean,
intercept_prior_standard_deviation,
intercept_proposal_standard_deviations, coefficient_prior_mean,
coefficient_prior_standard_deviation,
coefficient_proposal_standard_deviations,
latent_position_prior_mean, latent_position_prior_standard_deviation,
latent_position_proposal_standard_deviations, target_accept_rate, tollerance,
update_size, seed, iterations, metropolis_iterations, total_number_of_tokens,
iterations_before_t_i_p_updates, update_t_i_p_every_x_iterations,
perform_adaptive_metropolis must be provided.}

\item{Test_RDirichlet}{If TRUE, then alpha_m must be provided.}

\item{Test_Sample_Token_Topics_From_Generative_Process}{If TRUE, then optional
arguments author_indexes, token_topic_assignments, token_word_types, alpha_m,
beta_n, random_numbers, and resample_word_types must be provided.}

\item{seed}{An integerg.}

\item{distribution}{A log-space vector}

\item{intercepts}{A vector.}

\item{coefficients}{A matrix.}

\item{latent_positions}{An array.}

\item{document_sender}{An integer.}

\item{document_recipient}{An integer.}

\item{current_covariates}{A vector}

\item{interaction_pattern}{An integer}

\item{using_coefficients}{A boolean.}

\item{edge_probabilities}{An array}

\item{tokens_in_document}{An integer}

\item{current_token_topic_assignment}{An integer}

\item{current_document_topic_counts}{A vector.}

\item{leave_out_current_token}{A logical}

\item{topic_interaction_patterns}{a vector}

\item{leave_out_topic}{An integer.}

\item{intercept_prior_mean}{A double.}

\item{intercept_prior_standard_deviation}{A double.}

\item{intercept_proposal_standard_deviations}{A vector.}

\item{coefficient_prior_mean}{A double.}

\item{coefficient_prior_standard_deviation}{A double.}

\item{coefficient_proposal_standard_deviations}{A vector.}

\item{latent_position_prior_mean}{A double.}

\item{latent_position_prior_standard_deviation}{A double.}

\item{latent_position_proposal_standard_deviations}{A vector.}

\item{topic}{An integer.}

\item{document_edge_values}{A vector.}

\item{word_type_topic_counts}{A matrix}

\item{topic_token_counts}{A matrix.}

\item{current_word_type}{An integer.}

\item{alpha_m}{A vector}

\item{beta_n}{A vector}

\item{beta}{Is equal to sum(beta_n)}

\item{rand_num}{An integer}

\item{author_indexes}{A vector}

\item{document_edge_matrix}{A matrix}

\item{document_topic_counts}{A matrix}

\item{token_topic_assignments}{A List}

\item{token_word_types}{A List}

\item{covariates}{An array}

\item{random_numbers}{A vector}

\item{accept_rates}{A vector.}

\item{target_accept_rate}{A double.}

\item{tollerance}{A dobule.}

\item{update_size}{A double.}

\item{iterations}{An integer.}

\item{metropolis_iterations}{An integer.}

\item{total_number_of_tokens}{An integer.}

\item{iterations_before_t_i_p_updates}{An integer.}

\item{update_t_i_p_every_x_iterations}{An integer.}

\item{perform_adaptive_metropolis}{A boolean.}

\item{resample_word_types}{A boolean.}

\item{slice_sample_alpha_m}{An integer (normally a boolean but you can set
the value directly (normally 5, set to a negative number to turn off)),}

\item{slice_sample_step_size}{An integer.}

\item{parallel}{A boolean.}

\item{use_cached_token_topic_distribution}{A boolean.}

\item{cached_token_topic_distribution}{A vector with legth equal to the number
of topics.}
}
\value{
Whatever is returned by the internal function being tested
}
\description{
Allows the user to interface with the individual C++ functions.
}

